feat: 起動時にモーダル送信時のスクリプトを自動読み込み、起動時のログの改善
