+++ .vscode/settings.json
@@ -1,6 +1,7 @@
 {
   "cSpell.words": [
     "addproject",
+    "chelluoxou",
     "deploycommands",
     "Lubba",
     "Pongoe",
