+++ src/deploy-commands.js
@@ -46,7 +46,7 @@ for (const filePath of commandFiles) {
     commands.push(command.data.toJSON())
   } else {
     console.log(
-      `[WARNING] The command at ${filePath} is missing a required "data" or "execute" property.`
+      `⚠️ The command at ${filePath} is missing a required "data" or "execute" property.`
     )
   }
 }
@@ -54,7 +54,7 @@ for (const filePath of commandFiles) {
 ;(async () => {
   try {
     console.log(
-      `Started refreshing ${commands.length} application (/) commands.`
+      `✅ Started refreshing ${commands.length} application (/) commands.`
     )
 
     // The put method is used to fully refresh all commands in the guild with the current set
@@ -64,7 +64,7 @@ for (const filePath of commandFiles) {
     )
 
     console.log(
-      `Successfully reloaded ${data.length} application (/) commands.`
+      `✅ Successfully reloaded ${data.length} application (/) commands.`
     )
   } catch (error) {
     // And of course, make sure you catch and log any errors!
