+++ src/index.ts
@@ -7,14 +7,12 @@ import {
   Message,
 } from 'discord.js'
 import 'dotenv/config'
+import yargs from 'yargs'
 
-import commands from './command'
-import createProject from './utils/create-project'
 import { fetchProjects } from './adapters/notion-adapter'
 import { Projects } from './models/projects'
-import { editProject } from './interactions/modal/edit-project'
-
 import { loadCommands } from './command'
+import loadModalModules from './modal'
 
 const BOT_VERSION = '1.1.0'
 
@@ -33,6 +31,7 @@ const args = yargs
   .parseSync()
 
 const DEPLOY_COMMANDS_MODE = args['deploy-commands']
+let modalModules: Map<string, (interaction: any) => Promise<void>> = new Map()
 
 type ExecuteCallback = (interaction: any) => Promise<void>
 class MyClient extends Client {
@@ -53,18 +52,25 @@ const client = new MyClient({
 })
 
 client.once(Events.ClientReady, async (c: Client) => {
-  console.log('Ready! Logged in as ' + c.user?.tag)
+  console.log('✅ Logged in as ' + c.user?.tag + '\n')
 
-  console.log('Loading commands...')
+  console.log('⏳ Loading command modules...')
   client.commands = await loadCommands()
+  console.log('✅ Command modules loaded successfully.\n')
 
-  console.log('Fetching projects...')
+  console.log('⏳ Loading modal modules...')
+  modalModules = await loadModalModules()
+  console.log('✅ Modal modules loaded successfully.\n')
+
+  console.log('⏳ Fetching projects...')
   await guildProjectsCache.fetchProjects()
   console.log(
-    'Fetched projects successfully.',
+    '✅ Fetched projects successfully.',
     guildProjectsCache.projects.length,
-    'projects found.'
+    'projects found.\n'
   )
+
+  console.log('🚀 Spacecruiser is ready to launch!\n')
 })
 
 client.on(Events.InteractionCreate, async (interaction: BaseInteraction) => {
@@ -130,27 +136,19 @@ client.on(Events.InteractionCreate, async (interaction: BaseInteraction) => {
   }
 })
 
-// respond to modal submit
+// モーダル送信時の処理
 client.on(Events.InteractionCreate, async (interaction) => {
   if (!interaction.isModalSubmit()) return
+  const callback = modalModules.get(interaction.customId)
 
-  // add project command of ./commands/addproject.ts
-  if (interaction.customId === 'addprojectModal') {
-    const projectId = interaction.fields.getTextInputValue('projectIdInput')
-
-    const message = await interaction.reply({
-      content: 'Creating new project...',
+  if (callback) {
+    await callback(interaction)
+  } else {
+    console.log('No callback found for modal ID: ' + interaction.customId)
+    await interaction.reply({
+      content: '❌ An error occurred while processing the modal submission.',
+      ephemeral: true,
     })
-
-    createProject({
-      projectId: projectId,
-      interaction: message,
-    })
-  }
-
-  if (interaction.customId === 'editProjectModal') {
-    console.log('interaction:', interaction)
-    await editProject(interaction)
   }
 })
 
@@ -164,12 +162,10 @@ client.on(Events.MessageCreate, async (message: Message) => {
 })
 
 if (DEPLOY_COMMANDS_MODE) {
-  (async () => {
-    console.log('Deploying commands...')
-    await require('./deploy-commands')
-  })()
+  console.log('⏳ Deploying commands...')
+  import('./deploy-commands')
 } else {
-  console.log('Booting spacecruiser...')
+  console.log('⏳ Booting spacecruiser...')
   client.login(process.env.DISCORD_TOKEN)
 }
